Finished edit comment feature

+++ app/api/comment_routes.py
@@ -1,4 +1,5 @@
 from flask import Blueprint, request
+from app.api.auth_routes import validation_errors_to_error_messages
 from app.models import Comment, db
 from flask_login import login_required, current_user
 from app.forms import CommentForm
@@ -9,8 +10,31 @@ comment_routes = Blueprint('comments', __name__)
 @login_required
 def delete_comment(comment_id):
     comment_to_delete = Comment.query.filter(Comment.id == comment_id).first()
+    if comment_to_delete.user_id != current_user.id:
+        return {
+            "error": "You are not authorized to delete this comment"
+        }
     db.session.delete(comment_to_delete)
     db.session.commit()
     return {
             "comment": comment_to_delete.to_dict()
         }
+
+@comment_routes.route('/<int:comment_id>', methods=['PATCH'])
+@login_required
+def edit_comment(comment_id):
+    comment = CommentForm()
+    comment['csrf_token'].data = request.cookies['csrf_token']
+    if comment.validate_on_submit():
+        commentToUpdate = Comment.query.filter(Comment.id == comment_id).first()
+        if commentToUpdate.user_id != current_user.id:
+            return {
+                "error": "You are not authorized to edit this comment"
+            }
+        commentToUpdate.description = comment.data['description']
+        db.session.add(commentToUpdate)
+        db.session.commit()
+        return {
+            "message": "successful"
+        }
+    return {'errors': validation_errors_to_error_messages(comment.errors)}, 401
